requirements.txt
asgiref==3.3.1
autopep8==1.5.7
certifi==2020.12.5
charset-normalizer==2.0.3
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.11.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-gis==0.17
GDAL==3.0.4
idna==3.2
importlib-metadata==3.4.0
Markdown==3.3.3
pip-autoremove==0.9.1
psycopg2==2.8.6
pycodestyle==2.7.0
pygeoif==0.7
pyproj==3.0.0.post1
pyshp==2.1.3
python-dotenv==0.19.0
pytz==2020.5
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.6
zipp==3.4.0
